pyproject.toml
[tool.poetry]
name = "presentation"
version = "0.1.0"
description = ""
authors = ["theovald <gjoegjoe@outlook.com>"]

[tool.poetry.dependencies]
python =  ">=3.9,<3.9.7"
streamlit = "^1.12.1"
Pillow = "^9.2.0"
numpy = "^1.23.2"
opencv-python = "^4.6.0"
torch = "^1.12.1"
torchvision = "^0.13.1"
torchaudio = "^0.12.1"
Cython = "^0.29.32"
scipy = "^1.9.1"
matplotlib = "^3.5.3"
requests = "^2.28.1"
graphviz = "^0.20.1"

[tool.poetry.dev-dependencies]
black = "^22.6.0"   
isort = "^5.10.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
